[statics] Fix global error about default modules
This was caused by the query of queryMod requiring an M edge,
but that is obviously not present when querying to see if there are two
declarations of the same name in the scope that was just declared in.
To fix this, the query filter is adjusted to also allow epsilon edges.