[gzb2gzbc,gzb/trans] Implement Raw and Ref transformations
This required some major changes to the way references are resolved in
conjunction with the question of using NSID vs NSIDi.
These changes ensure that, at least for the purposes of Ref, all .ref
AST properties are properly referring to NSIDs (which in turn should have
associated AST annos with proper reference info), instead of the fictive
NSIDi nodes.