[gazebo,gzb2gzbc] Fix reference to members imported (e.g. from stdlib)
Because imported members are declared in the current module,
their absolute names would be generated relative to that module.
This is not correct, obviously. To correct this behavior,
the full absolute name is now included in the scope decl
for a member.
Ideally this would be modelled purely using scopes,
but with the current setup with types/members themselves
not yet being scopes, it is not possible.